# api/openapi.yaml
openapi: 3.0.3
info:
  title: CompAInion API
  version: 0.1.0
  description: |
    REST backend for CompAInion: a Hermes session acting as "companion"
    dispatches plans/tasks to detached Hermes workers and serves the user one
    interruption at a time.

    Conventions:
    - Times are RFC3339 UTC strings.
    - JSON-valued columns (`handle`, `scope`, `acceptance_checks`, `evidence`,
      `payload`, `outcome`) are typed as their natural JSON shape here.
    - Optimistic concurrency: plans/steps/tasks carry `version`; PATCH takes
      `If-Match: "<version>"` (missing -> 428, stale -> 412). Responses carry an
      `ETag: "<version>"` header.
    - Every creating `POST` accepts `Idempotency-Key`; replay returns the cached
      response with `Idempotent-Replayed: true`, same key + different body -> 422.
    - Lists are cursor paginated: `?cursor=<last id>&limit=<n>` (default 50, max
      200), envelope `{items, next_cursor}`.
    - Errors are RFC 7807 `application/problem+json` (400/404/409/412/422/428).
    - Step/task `status` is the stored status; `ready` is a derived boolean
      (pending + all deps done). A `ready` value is exposed ONLY via this flag.
servers:
  - url: http://127.0.0.1:7777
...
paths:
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      description: Filters `role`, `status`, `parent_id`, `workstream_id`.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - { name: role, in: query, schema: { $ref: '#/components/schemas/AgentRole' } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/AgentStatus' } }
        - { name: parent_id, in: query, schema: { type: string } }
        - { name: workstream_id, in: query, schema: { type: string } }
      responses:
        '200':
          description: Paginated agents
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Agent' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      operationId: registerAgent
      tags: [agents]
      description: Idempotent on a supplied `id` if given.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentInput' }
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        default: { $ref: '#/components/responses/Problem' }
  /agents/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getAgent
      tags: [agents]
      responses:
        '200':
          description: The agent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        '404': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
    patch:
      operationId: updateAgent
      tags: [agents]
      description: Status changes go through the agent state machine; nothing
        leaves `finished`/`failed` (terminal). The backend never changes status
        by itself (a stale heartbeat does not alter it).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
  /agents/{id}/heartbeat:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      operationId: agentHeartbeat
      tags: [agents]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204': { description: Recorded (sets `last_heartbeat_at`) }
        '404': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
...
  /interruptions:
    get:
      operationId: listInterruptions
      tags: [interruptions]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - { name: status, in: query, schema: { $ref: '#/components/schemas/InterruptionStatus' } }
        - { name: workstream_id, in: query, schema: { type: string } }
        - { name: topic, in: query, schema: { type: string } }
      responses:
        '200':
          description: Paginated interruptions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InterruptionPage' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      operationId: createInterruption
      tags: [interruptions]
      description: Validates the questions; inserts interruption + questions +
        suggestions in one transaction; status `open`.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InterruptionInput' }
            example:
              workstream_id: 01JABCWORKSTREAM000000
              plan_id: 01JABCPLAN000000000000000
              topic: plan-approval
              kind: approval
              priority: high
              digest: Plan "Ship search" needs approval before execution.
              blocking: true
              questions:
                - position: 1
                  text: Approve the plan?
                  answer_type: confirm
                  suggestions:
                    - label: Approve
                      rationale: All acceptance checks are testable.
                      recommended: true
                    - label: Revise first
                - position: 2
                  text: Who should orchestrate execution?
                  answer_type: choice
                  suggestions:
                    - label: Orchestrator agent A
                      recommended: true
                    - label: Keep it manual
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Interruption' }
        default: { $ref: '#/components/responses/Problem' }
  /interruptions/next:
    get:
      operationId: nextInterruption
      tags: [interruptions]
      description: |
        Picks the single highest-priority `open` interruption (blocking first,
        then priority rank, then oldest) and marks it `presented`; calling
        again does NOT return the same one. Also returns `batch` - the ids of
        the other `open` interruptions with the same `topic`. Returns 204 when
        none is open. `wait=<seconds>` long-polls; 204 on timeout.
      parameters:
        - name: wait
          in: query
          description: Long-poll up to this many seconds.
          schema: { type: integer, minimum: 0, maximum: 60 }
      responses:
        '200':
          description: One interruption (now `presented`) with `batch` ids
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Interruption' }
        '204':
          description: No open interruption (immediately or after `wait`)
        default: { $ref: '#/components/responses/Problem' }
  /interruptions/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getInterruption
      tags: [interruptions]
      description: Expanded with questions, suggestions and answers.
      responses:
        '200':
          description: The interruption
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Interruption' }
        '404': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
    delete:
      operationId: deleteInterruption
      tags: [interruptions]
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
  /interruptions/{id}/answers:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      operationId: answerInterruption
      tags: [interruptions]
      description: |
        Adds one answer. `mode` `suggestion`/`suggestion_with_comment` require a
        `suggestion_id` belonging to that question; `free`/`needs_details` take
        `text`. The user may use suggestion/suggestion_with_comment/free/
        needs_details; the companion may use pushback/free (clarification).
        With `final=true` the question becomes `answered`; when every question
        is `answered|skipped` the interruption becomes `answered` and
        `answered_at` is set.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnswerInput' }
      responses:
        '200':
          description: Answer recorded; returns the interruption
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Interruption' }
        '404': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
  /interruptions/{id}/dismiss:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      operationId: dismissInterruption
      tags: [interruptions]
      description: 'open|presented -> dismissed.'
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Dismissed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Interruption' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
  /interruptions/{id}/reopen:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      operationId: reopenInterruption
      tags: [interruptions]
      description: Re-queues the interruption to `open` so a companion whose
        context was cleared can re-present it.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Reopened
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Interruption' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor (the last id of the previous page).
      schema: { type: string }
    Limit:
      name: limit
      in: query
      description: Page size (default 50, max 200 for resources; 100/500 for events).
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: Client-supplied key caching the response of this creating
        POST. Same key + different body -> 422.
      schema: { type: string }
  responses:
    Problem:
      description: RFC 7807 problem details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: RFC 7807 problem details. `type` maps the domain error kind
        (not_found, conflict, invalid, unprocessable, precondition_failed,
        precondition_required).
      required: [type, title, status]
      properties:
        type: { type: string, example: not_found }
        title: { type: string, example: Resource not found }
        status: { type: integer, example: 404 }
        detail: { type: string, example: plan 01J... not found }
        instance: { type: string }
...
    AgentRole:
      type: string
      enum: [companion, orchestrator, worker]
    AgentStatus:
      type: string
      enum: [starting, running, waiting, idle, lost, finished, failed]
    Agent:
      type: object
      required: [id, role, status, created_at, updated_at]
      properties:
        id: { type: string }
        role: { $ref: '#/components/schemas/AgentRole' }
        parent_id: { type: string }
        harness: { type: string }
        handle:
          type: object
          additionalProperties: true
          description: Harness-specific handle (e.g. tmux session info).
        status: { $ref: '#/components/schemas/AgentStatus' }
        workstream_id: { type: string }
        last_heartbeat_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AgentInput:
      type: object
      required: [role]
      properties:
        id:
          type: string
          description: Supply to register idempotently.
        role: { $ref: '#/components/schemas/AgentRole' }
        parent_id: { type: string }
        harness: { type: string }
        handle: { type: object, additionalProperties: true }
        status: { $ref: '#/components/schemas/AgentStatus' }
        workstream_id: { type: string }
...
    InterruptionKind:
      type: string
      enum: [decision, approval, clarification, finding, fyi]
    InterruptionPriority:
      type: string
      enum: [urgent, high, normal, low]
    InterruptionStatus:
      type: string
      enum: [open, presented, answered, dismissed, expired]
    QuestionAnswerType:
      type: string
      enum: [choice, multi_choice, free_text, confirm]
    QuestionStatus:
      type: string
      enum: [open, answered, skipped]
    AnswerAuthor:
      type: string
      enum: [user, companion]
    AnswerMode:
      type: string
      enum: [suggestion, suggestion_with_comment, free, needs_details, pushback]
    Suggestion:
      type: object
      required: [label, recommended]
      properties:
        id: { type: string }
        label: { type: string }
        rationale: { type: string }
        recommended: { type: boolean }
    Question:
      type: object
      required: [position, text, answer_type]
      properties:
        id: { type: string }
        position: { type: integer }
        text: { type: string }
        answer_type: { $ref: '#/components/schemas/QuestionAnswerType' }
        group: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/QuestionStatus' }
        suggestions:
          type: array
          items: { $ref: '#/components/schemas/Suggestion' }
        answers:
          type: array
          items: { $ref: '#/components/schemas/Answer' }
    Answer:
      type: object
      required: [id, question_id, author, mode, final, created_at]
      properties:
        id: { type: string }
        question_id: { type: string }
        author: { $ref: '#/components/schemas/AnswerAuthor' }
        mode: { $ref: '#/components/schemas/AnswerMode' }
        suggestion_id: { type: string, nullable: true }
        text: { type: string, nullable: true }
        final: { type: boolean }
        created_at: { type: string, format: date-time }
    AnswerInput:
      type: object
      required: [question_id, author, mode]
      properties:
        question_id: { type: string }
        author: { $ref: '#/components/schemas/AnswerAuthor' }
        mode: { $ref: '#/components/schemas/AnswerMode' }
        suggestion_id: { type: string }
        text: { type: string }
        final: { type: boolean, default: false }
    SuggestionInput:
      type: object
      required: [label]
      properties:
        label: { type: string }
        rationale: { type: string }
        recommended: { type: boolean, default: false }
    QuestionInput:
      type: object
      required: [position, text, answer_type]
      properties:
        position: { type: integer }
        text: { type: string }
        answer_type: { $ref: '#/components/schemas/QuestionAnswerType' }
        group: { type: string }
        suggestions:
          type: array
          items: { $ref: '#/components/schemas/SuggestionInput' }
    Interruption:
      type: object
      required: [id, topic, kind, priority, digest, blocking, status,
        created_at, updated_at]
      properties:
        id: { type: string }
        workstream_id: { type: string }
        raised_by_agent_id: { type: string }
        plan_id: { type: string, nullable: true }
        step_id: { type: string, nullable: true }
        task_id: { type: string, nullable: true }
        topic: { type: string }
        kind: { $ref: '#/components/schemas/InterruptionKind' }
        priority: { $ref: '#/components/schemas/InterruptionPriority' }
        digest: { type: string }
        blocking: { type: boolean }
        status: { $ref: '#/components/schemas/InterruptionStatus' }
        answered_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        questions:
          type: array
          items: { $ref: '#/components/schemas/Question' }
        batch:
          type: array
          description: Ids of the other `open` interruptions with the same
            topic (returned by GET /interruptions/next).
          items: { type: string }
    InterruptionInput:
      type: object
      required: [topic, kind, priority, digest]
      properties:
        workstream_id: { type: string }
        raised_by_agent_id: { type: string }
        plan_id: { type: string }
        step_id: { type: string }
        task_id: { type: string }
        topic: { type: string }
        kind: { $ref: '#/components/schemas/InterruptionKind' }
        priority: { $ref: '#/components/schemas/InterruptionPriority' }
        digest: { type: string }
        blocking: { type: boolean, default: true }
        questions:
          type: array
          items: { $ref: '#/components/schemas/QuestionInput' }
...
    Page:
      type: object
      description: Standard list envelope; `next_cursor` is null on the last page.
      required: [items]
      properties:
        next_cursor: { type: string, nullable: true }
...
    InterruptionPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Interruption' }
...